pyproject.toml
[project]
name = "claude-9-q-hack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52.0",
    "python-dotenv>=1.1.0",
    "flask>=3.1.3",
    "pyaudio>=0.2.14",
    "requests>=2.32.3",
    "speechrecognition>=3.16.0",
]
